crap just noticed that our colour mods should be unsigned long long

+++ drivers/video/msm/smart_mtp_se6e8fa.c
@@ -89,7 +89,7 @@ static char VT_300CD_R;
 static char VT_300CD_G;
 static char VT_300CD_B;
 
-static long color_mods[5][9] = {
+static unsigned long long color_mods[5][9] = {
 	{  0,  0,  5, -18, -16, -10, -5, 0,  0 },
 	{  0,  0,  2,  -9,  -8,  -5, -2, 0,  0 },
 	{  0,  0,  0,   0,   0,   0,  0, 0,  0 },
